chore: clean up all clippy warnings (idiomatic pass)
Zero-behavior-change lint cleanup so `cargo clippy --all-targets`
is warning-free and stays that way:

- useless format! → to_string() (fetch.rs X-Robots-Tag suggestions)
- manual range checks → RangeInclusive::contains (design_slop, voice)
- map(..).flatten() → and_then (freshness visible-label)
- redundant closure → and_then(host_of) (links)
- manual prefix strip → strip_prefix (markdown frontmatter)
- redundant f32→f32 cast (copy_precision density)
- collapse identical if/else-if arms into one guard (metatext verdict)
- document the three cohesive-arg functions with #[allow(too_many_arguments)]
  rather than inventing artificial params structs

Tests unchanged and green (41 passing).

+++ src/audit/copy_precision.rs
@@ -236,7 +236,7 @@ pub fn extract(body_text: &str) -> CopyPrecision {
     );
     densities.insert(
         "concrete_per_100_words",
-        (concrete_per_100 * 100.0).round() as f32 / 100.0,
+        (concrete_per_100 * 100.0).round() / 100.0,
     );
 
     CopyPrecision {
